% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_selection.R
\name{glimpse_selection}
\alias{glimpse_selection}
\alias{str_selection}
\title{Glimpse at Object Selected in the Editor}
\usage{
glimpse_selection()

str_selection()
}
\description{
Glimpse at Object Selected in the Editor
}

