% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_selection.R
\name{glimpse_selection}
\alias{glimpse_selection}
\alias{str_selection}
\title{Glimpse at Object Selected in the Editor}
\usage{
glimpse_selection()

str_selection()
}
\description{
Select an object name in the RStudio code editor. Executing these addins will
run \code{dplyr::glimpse} or \code{utils::str} on that object showing its
content in a compact way.
}

